moves diagram component into groups for reuse
