+++ src/groups/RouteNetworkDiagram/AddInnerSpanStructure/AddInnerSpanStructure.tsx
