+++ src/groups/RouteNetworkDiagram/AddInnerSpanStructure/AddInnerSpanStructureGql.ts
