+++ src/groups/RouteNetworkDiagram/AddInnerSpanStructure/index.ts
