+++ src/groups/RouteNetworkDiagram/EditSpanEquipment/EditSpanEquipment.tsx
