+++ src/groups/RouteNetworkDiagram/EditSpanEquipment/EditspanEquipmentGql.ts
