+++ src/groups/RouteNetworkDiagram/EditSpanEquipment/index.ts
