+++ src/groups/RouteNetworkDiagram/IdentifyFeaturePage.scss
@@ -1,0 +1,40 @@
+.identify-feature-page {
+  width: 100%;
+
+  .feature-information-container {
+    display: flex;
+    padding: 20px 0px;
+    justify-content: center;
+    border-bottom: 1px solid #e3e3e3;
+
+    .feature-informations {
+      display: flex;
+      flex-wrap: wrap;
+      max-width: 1000px;
+      justify-content: space-between;
+      width: 100%;
+      font-size: 20px;
+    }
+  }
+
+  .feature-details {
+    border-top: 1px solid #e3e3e3;
+    padding: 20px;
+
+    .feature-details-container {
+      display: flex;
+      margin: 0px auto;
+      max-width: 1000px;
+
+      .feature-details-info {
+        flex: 1;
+      }
+
+      .feature-details-actions {
+        .action-button {
+          margin-bottom: 10px;
+        }
+      }
+    }
+  }
+}
