+++ src/groups/RouteNetworkDiagram/SchematicDiagram/SchematicDiagram.scss
