+++ src/groups/RouteNetworkDiagram/SchematicDiagram/SchematicDiagram.tsx
