+++ src/groups/RouteNetworkDiagram/SchematicDiagram/colors.ts
