+++ src/groups/RouteNetworkDiagram/SchematicDiagram/diagramLayer.ts
