+++ src/groups/RouteNetworkDiagram/SchematicDiagram/index.ts
