+++ src/groups/RouteNetworkDiagram/SpanEquipmentDetails/SpanEquipmentDetails.scss
