+++ src/groups/RouteNetworkDiagram/SpanEquipmentDetails/SpanEquipmentDetails.tsx
