+++ src/groups/RouteNetworkDiagram/SpanEquipmentDetails/SpanEquipmentDetailsGql.ts
