+++ src/groups/RouteNetworkDiagram/SpanEquipmentDetails/index.ts
